Create App.js
This is running React-Router-dom v6, so if you have an older model the Routes won't be the same name and the parameters will require additional info. For the App.js file I have it set up where it holds the pages in components which makes it easier to read and more organized.