Create client
This is the initial app which consists of all the components being outputted depending on the route. Each route is serving as a page. With the initial it's important to remember that there is no protection to reach the blog form post. I will be adding the code to have only authenticated user's to have access to the back-end.