Create App.js
This is the initial App.js where the server only holds the database files needed to run this program. App.js serves as the page holder and route controller. Each route has a different component which is a different page. Note that this version has no protection and anyone can access the blog form. The initial version doesn't require any authentication and the server folder will only contain the blog database using MongoDB.