Create index.js
Index.js is the file that will overall render all of the components into index.html. The App component is wrapped around the BlogsContextProvider component which basically allows global usage of context in the overall site. A context is a global custom hook that can be used in any files. For our hook we allow the app to access our database.