Create Navbar.js
My navbar in it's initial form doesn't work currently for the mobile part, but it's a simple fix and this is the initial file, so I don't really think it's a big deal. The initial files are meant to show my thought process of how I went about creating this web app for anyone to start out and just use. This also gives the opportunity to be creative with creating the navbar.