Add files via upload
other important files. I had all these files in a folder called functions, the purpose is that firebase doesn't recognize the back end server if it's not named functions. Firebase.json is a file for firebase config and the routes is for user and blog routes to connect with the data it holds in Mongo.